upload file when closed
Current solution can't handle updating/editing a file cause it only appends to the blob. So, it can only write sequences of blocks, but we have default_storage.save(content) for that, which works better cause it avoids creating a disk file. There is a way to update a blob (see get_block_list) but since the file position can be moved back and forth, it gets quite complicated to keep the blocks in sync (there may be missing blocks between two position, etc). Until then, this seems like the better way.